Added class "text". This may break built-in font size.
Some books use "text" as the default text field, on which vitalsource's font size tool does work properly. This change to the script now overrides vitalsource's font size tool. (I was unhappy because the smallest text option was slightly too small and the next size up was way too large, so I decided the built-in tool wasn't worth using even when it was functioning as intended)